attempt 2: setup bot to procedure and call it after all all services started
